make new class for user input
